Bug fix, better plots, continuous trades
Fixed indexing and buy/sell indicator when lumping.
Plots now indicate the size of the trade volume.
Continuous trades now you can look at the prevailing bid/ask as the reference, or you can use the most recent trade as the price reference. The latter will have the buy/sell value 0, because it is not clear who crosses the spread.